requirements.txt
aws-cdk-lib
aws-cdk-constructs
aws-cdk.aws-lambda-python-alpha
pydantic
pydantic-settings